refactor: move socket (#92)
* feat: impl Iterator for Transport

* feat: iterate over packets instead

* feat: prepare for iter()

* refactor: moved client side logic

* feat: qol

* fix: broken format!

* feat: fix socketio test

* feat: add tls_connector dynamic secure

* feat: less need to clone

* feat: move socket
