+++ src/engineio/client/mod.rs
@@ -1,0 +1,3 @@
+mod socket;
+pub(crate) use socket::Iter;
+pub use {socket::Iter as SocketIter, socket::Socket, socket::SocketBuilder};
