+++ src/engineio/transport.rs
@@ -72,3 +72,5 @@ impl std::fmt::Debug for dyn Transport {
         f.write_fmt(format_args!("Transport(base_url: {:?})", self.base_url(),))
     }
 }
+
+//TODO: add iter() implementation
