+++ src/socketio/client/socket.rs
@@ -1,0 +1,487 @@
+pub use super::super::{event::Event, payload::Payload};
+use crate::socketio::packet::Packet as SocketPacket;
+use native_tls::TlsConnector;
+pub use reqwest::header::{HeaderMap, HeaderValue, IntoHeaderName};
+
+use crate::error::Result;
+use std::{time::Duration, vec};
+
+use crate::socketio::socket::Socket as InnerSocket;
+
+/// A socket which handles communication with the server. It's initialized with
+/// a specific address as well as an optional namespace to connect to. If `None`
+/// is given the server will connect to the default namespace `"/"`.
+#[derive(Debug, Clone)]
+pub struct Socket {
+    /// The inner socket client to delegate the methods to.
+    // TODO: Make this private
+    pub socket: InnerSocket,
+}
+
+type SocketCallback = dyn FnMut(Payload, Socket) + 'static + Sync + Send;
+
+/// A builder class for a `socket.io` socket. This handles setting up the client and
+/// configuring the callback, the namespace and metadata of the socket. If no
+/// namespace is specified, the default namespace `/` is taken. The `connect` method
+/// acts the `build` method and returns a connected [`Socket`].
+pub struct SocketBuilder {
+    address: String,
+    on: Option<Vec<(Event, Box<SocketCallback>)>>,
+    namespace: Option<String>,
+    tls_config: Option<TlsConnector>,
+    opening_headers: Option<HeaderMap>,
+}
+
+impl SocketBuilder {
+    /// Create as client builder from a URL. URLs must be in the form
+    /// `[ws or wss or http or https]://[domain]:[port]/[path]`. The
+    /// path of the URL is optional and if no port is given, port 80
+    /// will be used.
+    /// # Example
+    /// ```rust
+    /// use rust_socketio::{SocketBuilder, Payload};
+    /// use serde_json::json;
+    ///
+    ///
+    /// let callback = |payload: Payload, _| {
+    ///            match payload {
+    ///                Payload::String(str) => println!("Received: {}", str),
+    ///                Payload::Binary(bin_data) => println!("Received bytes: {:#?}", bin_data),
+    ///            }
+    /// };
+    ///
+    /// let mut socket = SocketBuilder::new("http://localhost:4200")
+    ///     .namespace("/admin")
+    ///     .on("test", callback)
+    ///     .connect()
+    ///     .expect("error while connecting");
+    ///
+    /// // use the socket
+    /// let json_payload = json!({"token": 123});
+    ///
+    /// let result = socket.emit("foo", json_payload);
+    ///
+    /// assert!(result.is_ok());
+    /// ```
+    pub fn new<T: Into<String>>(address: T) -> Self {
+        Self {
+            address: address.into(),
+            on: None,
+            namespace: None,
+            tls_config: None,
+            opening_headers: None,
+        }
+    }
+
+    /// Sets the target namespace of the client. The namespace should start
+    /// with a leading `/`. Valid examples are e.g. `/admin`, `/foo`.
+    pub fn namespace<T: Into<String>>(mut self, namespace: T) -> Self {
+        let mut nsp = namespace.into();
+        if !nsp.starts_with('/') {
+            nsp = "/".to_owned() + &nsp;
+        }
+        self.namespace = Some(nsp);
+        self
+    }
+
+    /// Registers a new callback for a certain [`socketio::event::Event`]. The event could either be
+    /// one of the common events like `message`, `error`, `connect`, `close` or a custom
+    /// event defined by a string, e.g. `onPayment` or `foo`.
+    /// # Example
+    /// ```rust
+    /// use rust_socketio::{SocketBuilder, Payload};
+    ///
+    /// let callback = |payload: Payload, _| {
+    ///            match payload {
+    ///                Payload::String(str) => println!("Received: {}", str),
+    ///                Payload::Binary(bin_data) => println!("Received bytes: {:#?}", bin_data),
+    ///            }
+    /// };
+    ///
+    /// let socket = SocketBuilder::new("http://localhost:4200/")
+    ///     .namespace("/admin")
+    ///     .on("test", callback)
+    ///     .on("error", |err, _| eprintln!("Error: {:#?}", err))
+    ///     .connect();
+    ///
+    /// ```
+    pub fn on<F>(mut self, event: &str, callback: F) -> Self
+    where
+        F: FnMut(Payload, Socket) + 'static + Sync + Send,
+    {
+        match self.on {
+            Some(ref mut vector) => vector.push((event.into(), Box::new(callback))),
+            None => self.on = Some(vec![(event.into(), Box::new(callback))]),
+        }
+        self
+    }
+
+    /// Uses a preconfigured TLS connector for secure cummunication. This configures
+    /// both the `polling` as well as the `websocket` transport type.
+    /// # Example
+    /// ```rust
+    /// use rust_socketio::{SocketBuilder, Payload};
+    /// use native_tls::TlsConnector;
+    ///
+    /// let tls_connector =  TlsConnector::builder()
+    ///            .use_sni(true)
+    ///            .build()
+    ///            .expect("Found illegal configuration");
+    ///
+    /// let socket = SocketBuilder::new("http://localhost:4200/")
+    ///     .namespace("/admin")
+    ///     .on("error", |err, _| eprintln!("Error: {:#?}", err))
+    ///     .tls_config(tls_connector)
+    ///     .connect();
+    ///
+    /// ```
+    pub fn tls_config(mut self, tls_config: TlsConnector) -> Self {
+        self.tls_config = Some(tls_config);
+        self
+    }
+
+    /// Sets custom http headers for the opening request. The headers will be passed to the underlying
+    /// transport type (either websockets or polling) and then get passed with every request thats made.
+    /// via the transport layer.
+    /// # Example
+    /// ```rust
+    /// use rust_socketio::{SocketBuilder, Payload};
+    /// use reqwest::header::{ACCEPT_ENCODING};
+    ///
+    ///
+    /// let socket = SocketBuilder::new("http://localhost:4200/")
+    ///     .namespace("/admin")
+    ///     .on("error", |err, _| eprintln!("Error: {:#?}", err))
+    ///     .opening_header(ACCEPT_ENCODING, "application/json".parse().unwrap())
+    ///     .connect();
+    ///
+    /// ```
+    pub fn opening_header<K: IntoHeaderName>(mut self, key: K, val: HeaderValue) -> Self {
+        match self.opening_headers {
+            Some(ref mut map) => {
+                map.insert(key, val);
+            }
+            None => {
+                let mut map = HeaderMap::new();
+                map.insert(key, val);
+                self.opening_headers = Some(map);
+            }
+        }
+        self
+    }
+
+    /// Connects the socket to a certain endpoint. This returns a connected
+    /// [`Socket`] instance. This method returns an [`std::result::Result::Err`]
+    /// value if something goes wrong during connection.
+    /// # Example
+    /// ```rust
+    /// use rust_socketio::{SocketBuilder, Payload};
+    /// use serde_json::json;
+    ///
+    ///
+    /// let mut socket = SocketBuilder::new("http://localhost:4200/")
+    ///     .namespace("/admin")
+    ///     .on("error", |err, _| eprintln!("Socket error!: {:#?}", err))
+    ///     .connect()
+    ///     .expect("connection failed");
+    ///
+    /// // use the socket
+    /// let json_payload = json!({"token": 123});
+    ///
+    /// let result = socket.emit("foo", json_payload);
+    ///
+    /// assert!(result.is_ok());
+    /// ```
+    pub fn connect(self) -> Result<Socket> {
+        //TODO: Add check for empty path and set to /socket.io/
+        let mut socket = Socket::new(
+            self.address,
+            self.namespace,
+            self.tls_config,
+            self.opening_headers,
+        )?;
+        if let Some(callbacks) = self.on {
+            for (event, callback) in callbacks {
+                socket.on(event, Box::new(callback)).unwrap();
+            }
+        }
+        socket.connect()?;
+        Ok(socket)
+    }
+}
+
+impl Socket {
+    /// Creates a socket with a certain address to connect to as well as a
+    /// namespace. If `None` is passed in as namespace, the default namespace
+    /// `"/"` is taken.
+    /// ```
+    pub(crate) fn new<T: Into<String>>(
+        address: T,
+        namespace: Option<String>,
+        tls_config: Option<TlsConnector>,
+        opening_headers: Option<HeaderMap>,
+    ) -> Result<Self> {
+        Ok(Socket {
+            socket: InnerSocket::new(address, namespace, tls_config, opening_headers)?,
+        })
+    }
+
+    /// Registers a new callback for a certain event. This returns an
+    /// `Error::IllegalActionAfterOpen` error if the callback is registered
+    /// after a call to the `connect` method.
+    pub(crate) fn on<F>(&mut self, event: Event, callback: Box<F>) -> Result<()>
+    where
+        F: FnMut(Payload, Self) + 'static + Sync + Send,
+    {
+        self.socket.on(event, callback)
+    }
+
+    /// Connects the client to a server. Afterwards the `emit_*` methods can be
+    /// called to interact with the server. Attention: it's not allowed to add a
+    /// callback after a call to this method.
+    pub(crate) fn connect(&mut self) -> Result<()> {
+        self.socket.connect()
+    }
+
+    /// Sends a message to the server using the underlying `engine.io` protocol.
+    /// This message takes an event, which could either be one of the common
+    /// events like "message" or "error" or a custom event like "foo". But be
+    /// careful, the data string needs to be valid JSON. It's recommended to use
+    /// a library like `serde_json` to serialize the data properly.
+    ///
+    /// # Example
+    /// ```
+    /// use rust_socketio::{SocketBuilder, Socket, Payload};
+    /// use serde_json::json;
+    ///
+    /// let mut socket = SocketBuilder::new("http://localhost:4200/")
+    ///     .on("test", |payload: Payload, socket: Socket| {
+    ///         println!("Received: {:#?}", payload);
+    ///         socket.emit("test", json!({"hello": true})).expect("Server unreachable");
+    ///      })
+    ///     .connect()
+    ///     .expect("connection failed");
+    ///
+    /// let json_payload = json!({"token": 123});
+    ///
+    /// let result = socket.emit("foo", json_payload);
+    ///
+    /// assert!(result.is_ok());
+    /// ```
+    #[inline]
+    pub fn emit<E, D>(&self, event: E, data: D) -> Result<()>
+    where
+        E: Into<Event>,
+        D: Into<Payload>,
+    {
+        self.socket.emit(event.into(), data.into())
+    }
+
+    /// Disconnects this client from the server by sending a `socket.io` closing
+    /// packet.
+    /// # Example
+    /// ```rust
+    /// use rust_socketio::{SocketBuilder, Payload, Socket};
+    /// use serde_json::json;
+    ///
+    /// let mut socket = SocketBuilder::new("http://localhost:4200/")
+    ///     .on("test", |payload: Payload, socket: Socket| {
+    ///         println!("Received: {:#?}", payload);
+    ///         socket.emit("test", json!({"hello": true})).expect("Server unreachable");
+    ///      })
+    ///     .connect()
+    ///     .expect("connection failed");
+    ///
+    /// let json_payload = json!({"token": 123});
+    ///
+    /// socket.emit("foo", json_payload);
+    ///
+    /// // disconnect from the server
+    /// socket.disconnect();
+    ///
+    /// ```
+    pub fn disconnect(&mut self) -> Result<()> {
+        self.socket.disconnect()
+    }
+
+    /// Sends a message to the server but `alloc`s an `ack` to check whether the
+    /// server responded in a given timespan. This message takes an event, which
+    /// could either be one of the common events like "message" or "error" or a
+    /// custom event like "foo", as well as a data parameter. But be careful,
+    /// in case you send a [`Payload::String`], the string needs to be valid JSON.
+    /// It's even recommended to use a library like serde_json to serialize the data properly.
+    /// It also requires a timeout `Duration` in which the client needs to answer.
+    /// If the ack is acked in the correct timespan, the specified callback is
+    /// called. The callback consumes a [`Payload`] which represents the data send
+    /// by the server.
+    ///
+    /// # Example
+    /// ```
+    /// use rust_socketio::{SocketBuilder, Payload};
+    /// use serde_json::json;
+    /// use std::time::Duration;
+    /// use std::thread::sleep;
+    ///
+    /// let mut socket = SocketBuilder::new("http://localhost:4200/")
+    ///     .on("foo", |payload: Payload, _| println!("Received: {:#?}", payload))
+    ///     .connect()
+    ///     .expect("connection failed");
+    ///
+    ///
+    ///
+    /// let ack_callback = |message: Payload, _| {
+    ///     match message {
+    ///         Payload::String(str) => println!("{}", str),
+    ///         Payload::Binary(bytes) => println!("Received bytes: {:#?}", bytes),
+    ///    }    
+    /// };
+    ///
+    /// let payload = json!({"token": 123});
+    /// socket.emit_with_ack("foo", payload, Duration::from_secs(2), ack_callback).unwrap();
+    ///
+    /// sleep(Duration::from_secs(2));
+    /// ```
+    #[inline]
+    pub fn emit_with_ack<F, E, D>(
+        &self,
+        event: E,
+        data: D,
+        timeout: Duration,
+        callback: F,
+    ) -> Result<()>
+    where
+        F: FnMut(Payload, Socket) + 'static + Send + Sync,
+        E: Into<Event>,
+        D: Into<Payload>,
+    {
+        self.socket
+            .emit_with_ack(event.into(), data.into(), timeout, callback)
+    }
+
+    pub fn iter(&self) -> Iter {
+        Iter {
+            socket_iter: self.socket.iter(),
+        }
+    }
+}
+
+pub struct Iter<'a> {
+    socket_iter: crate::socketio::socket::Iter<'a>,
+}
+
+impl<'a> Iterator for Iter<'a> {
+    type Item = Result<SocketPacket>;
+    fn next(&mut self) -> std::option::Option<<Self as std::iter::Iterator>::Item> {
+        self.socket_iter.next()
+    }
+}
+
+#[cfg(test)]
+mod test {
+
+    use std::thread::sleep;
+
+    use super::*;
+    use crate::socketio::payload::Payload;
+    use bytes::Bytes;
+    use native_tls::TlsConnector;
+    use reqwest::header::{ACCEPT_ENCODING, HOST};
+    use serde_json::json;
+    use std::time::Duration;
+
+    #[test]
+    fn socket_io_integration() -> Result<()> {
+        //TODO: check to make sure we are receiving packets rather than logging.
+        let url = crate::socketio::test::socket_io_server()?;
+
+        let mut socket = Socket::new(url, None, None, None)?;
+
+        let result = socket.on(
+            "test".into(),
+            Box::new(|msg, _| match msg {
+                Payload::String(str) => println!("Received string: {}", str),
+                Payload::Binary(bin) => println!("Received binary data: {:#?}", bin),
+            }),
+        );
+        assert!(result.is_ok());
+
+        let result = socket.connect();
+        assert!(result.is_ok());
+
+        let payload = json!({"token": 123});
+        let result = socket.emit("test", Payload::String(payload.to_string()));
+
+        assert!(result.is_ok());
+
+        let ack_callback = move |message: Payload, socket: Socket| {
+            let result = socket.emit(
+                "test",
+                Payload::String(json!({"got ack": true}).to_string()),
+            );
+            assert!(result.is_ok());
+
+            println!("Yehaa! My ack got acked?");
+            if let Payload::String(str) = message {
+                println!("Received string Ack");
+                println!("Ack data: {}", str);
+            }
+        };
+
+        let ack = socket.emit_with_ack(
+            "test",
+            Payload::String(payload.to_string()),
+            Duration::from_secs(2),
+            ack_callback,
+        );
+        assert!(ack.is_ok());
+
+        socket.disconnect().unwrap();
+        // assert!(socket.disconnect().is_ok());
+
+        sleep(Duration::from_secs(10));
+
+        Ok(())
+    }
+
+    #[test]
+    fn socket_io_builder_integration() -> Result<()> {
+        let url = crate::socketio::test::socket_io_server()?;
+
+        // test socket build logic
+        let socket_builder = SocketBuilder::new(url);
+
+        let tls_connector = TlsConnector::builder()
+            .use_sni(true)
+            .build()
+            .expect("Found illegal configuration");
+
+        let socket = socket_builder
+            .namespace("/")
+            .tls_config(tls_connector)
+            .opening_header(HOST, "localhost".parse().unwrap())
+            .opening_header(ACCEPT_ENCODING, "application/json".parse().unwrap())
+            .on("test", |str, _| println!("Received: {:#?}", str))
+            .on("message", |payload, _| println!("{:#?}", payload))
+            .connect();
+
+        assert!(socket.is_ok());
+
+        let socket = socket.unwrap();
+        assert!(socket.emit("message", json!("Hello World")).is_ok());
+
+        assert!(socket.emit("binary", Bytes::from_static(&[46, 88])).is_ok());
+
+        let ack_cb = |payload, _| {
+            println!("Yehaa the ack got acked");
+            println!("With data: {:#?}", payload);
+        };
+
+        assert!(socket
+            .emit_with_ack("binary", json!("pls ack"), Duration::from_secs(1), ack_cb,)
+            .is_ok());
+
+        sleep(Duration::from_secs(5));
+
+        Ok(())
+    }
+}
