+++ src/socketio/mod.rs
@@ -1,11 +1,11 @@
+pub mod client;
 /// Defines the events that could be sent or received.
 pub mod event;
-mod packet;
+pub(crate) mod packet;
 /// Defines the types of payload (binary or string), that
 /// could be sent or received.
 pub mod payload;
-pub mod socket;
-pub(crate) mod transport;
+pub(self) mod socket;
 
 #[cfg(test)]
 pub(crate) mod test {
@@ -15,6 +15,12 @@ pub(crate) mod test {
 
     pub(crate) fn socket_io_server() -> crate::error::Result<Url> {
         let url = std::env::var("SOCKET_IO_SERVER").unwrap_or_else(|_| SERVER_URL.to_owned());
-        Ok(Url::parse(&url)?)
+        let mut url = Url::parse(&url)?;
+
+        if url.path() == "/" {
+            url.set_path("/socket.io/");
+        }
+
+        Ok(url)
     }
 }
